### 2019-09-19 #1749 @kthyng
@BIPL-HORIE I think the paper isn't compiling because some of the ORCIDs haven't been filled in for co-authors.
