### 2019-09-21 #1749 @arokem
👋 @hiendn : would you be willing to review this paper for JOSS?
