### 2019-09-21 #1749 @arokem
👋 @dirmeier : would you be willing to review this paper for JOSS?
