### 2019-09-23 #1749 @dirmeier
Hey @arokem, I can certainly give it a try. But I think it would be good to have a second reviewer (and third) on this one. Also I don't have a windows PC and the tool doesn't seem to have a description for installing on Unix systems ([here](https://github.com/BIPL-HORIE/LIBSDNN/blob/master/doc/Command%20line%20tools%20usage.pdf))
