### 2019-10-05 #1785 @whedon
[REVIEW]: LIBSDNN: A Selective Desensitization Neural Network Library
**Submitting author:** @BIPL-HORIE (<a href="http://orcid.org/0000-0001-7320-5478">和正 堀江</a>)
**Repository:** <a href="https://github.com/BIPL-HORIE/LIBSDNN" target ="_blank">https://github.com/BIPL-HORIE/LIBSDNN</a>
**Version:** v1.0
