### 2019-10-10 #1785 @arokem
@BIPL-HORIE : looks like the pdf  is still not building because of mismatches between the references in the paper.md and paper.bib files for example `Weston:2011` vs. `Weston:2000`. Could you please make sure that they match?
