### 2019-10-14 b8dda94
Update paper.md
correct the citation.
