+++ doc/paper.md
@@ -41,7 +41,7 @@ To facilitate the implementation and application of SDNN, we developed a library
 
 Although deep learning has achieved great success in various fields, it generally requires a large amount of data and has a high computational cost; moreover, non-experts face difficulties using it to its full potential. In contrast, ``shallow learning``, in which only one (usually the last) layer is trained, requires a much lesser amount of data, has low computational cost, and is still useful for many application.
 
-A representative learning machine with shallow learning is a support vector machine (SVM), which is very popular for the classification of extracted features. However, the performance of an SVM depends sensitively on the type of kernel and the hyperparameters. Its performance can degrade without hyperparameter optimization, which requires considerable effort; the performance can also degrade when the input vector contains irrelevant features [@Weston:2011] or when regions with largely different properties (such as density of training samples and complexity of decision boundaries) coexist. These shortcomings are more prominent in the regression version of an SVM, that is, support vector regression (SVR).
+A representative learning machine with shallow learning is a support vector machine (SVM), which is very popular for the classification of extracted features. However, the performance of an SVM depends sensitively on the type of kernel and the hyperparameters. Its performance can degrade without hyperparameter optimization, which requires considerable effort; the performance can also degrade when the input vector contains irrelevant features [@Weston:2000] or when regions with largely different properties (such as density of training samples and complexity of decision boundaries) coexist. These shortcomings are more prominent in the regression version of an SVM, that is, support vector regression (SVR).
 
 Another type of shallow neural network that overcomes these problems is a selective desensitization neural network (SDNN). An SDNN is a parallel perceptron (PP) [@Auer:2008] with two fixed preprocessing layers, in which two operations, namely, ``pattern coding`` [@Tanno:2015] and ``selective desensitization`` [@Morita:2005] are performed instead of kernel transformation (Please refer to SDNN section for more details). The PP part consists of some simple perceptrons (SPs) and represents the value of a function as the sum of the outputs of the SPs. Although each SP is a linear binary classifier, it can form very complex decision boundaries in the original input space because of the two abovementioned operations.
 
@@ -79,7 +79,7 @@ We randomly chose 2000 points from the 1001×1001 lattice points except for the 
 
 Figure 1(c) shows the estimation result. We can see that the SDNN approximates the target function properly even in the blank areas, indicating a very high generalization ability. The approximation error (root mean squared error (RMSE)) was 0.0859. As for the computational cost of LIBSDNN, the training process took about 240 s and required 8 MB of memory resources.
 
-We compared the above performance with LIBSVM [@Chang:2001], which is one of the most popular and effective approximation libraries and can also be used for SVR. In this comparison, the radial basis function was used as the kernel of SVR and its parameters, (C, $\gamma$, p), were optimized by the grid search method and were set to ($2^5$, $2^5$, 0.005).
+We compared the above performance with LIBSVM [@Chang:2007], which is one of the most popular and effective approximation libraries and can also be used for SVR. In this comparison, the radial basis function was used as the kernel of SVR and its parameters, (C, $\gamma$, p), were optimized by the grid search method and were set to ($2^5$, $2^5$, 0.005).
 
 The RMSE for the LIBSVM was 0.2136, which is more than twice of that for LIBSDNN. In particular, the approximation error was markedly large in the center blank area and the upper right area where the target function had a high spatial frequency, as shown in Figure 1(d). LIBSVM also had higher computational cost (Table 1). For your information, the comparison experiment was run on a 64-bit machine with Intel Core i7-3770K CPU (3.50 GHz, 4-core), 8
 MB cache, and 32 GB main memory.
