### 2019-10-14 #1785 @dirmeier
You can check if the pdf builds via: @whedon generate pdf
