+++ doc/paper.md
@@ -49,7 +49,7 @@ An SDNN provides high expression and generalization abilities. For example, it c
 
 One of the important advantages of an SDNN is its robustness against irrelevant features or redundant input dimensions [@Kobayashi:2015;@Tanno:2017]. Although extra dimensions increase the computational cost (second-order of the number of input dimensions), they have only a slight influence on the performance. Accordingly, we can save the effort required for performing feature selection or dimensionality reduction. 
 
-Another characteristic advantage of an SDNN is that the hyperparameter setting is easy. This is because its structure and learning algorithm are rather simple, and there exist only a few hyperparameters. Moreover, although its expression ability depends on and increases with the number of neural elements, overfitting does not occur because redundant elements are virtually unused or play the same role as other elements. The other hyperparameters do not affect the performance much if their values satisfy certain conditions (see "Selective desensitization neural network.pdf" in /doc folder). Accordingly, extensive hyperparameter search is unnecessary and we can use fixed hyperparameters or at most attempt several settings. Furthermore, even if the regions with different properties and optimal hyperparameters are mixed, the performance does not degrade much for any of the regions.
+Another characteristic advantage of an SDNN is that the hyperparameter setting is easy. This is because its structure and learning algorithm are rather simple, and there exist only a few hyperparameters. Moreover, although its expression ability depends on and increases with the number of neural elements, overfitting does not occur because redundant elements are virtually unused or play the same role as other elements. The other hyperparameters do not affect the performance much if their values satisfy certain conditions (see "Selective desensitization neural network.pdf" in &frasl;doc folder). Accordingly, extensive hyperparameter search is unnecessary and we can use fixed hyperparameters or at most attempt several settings. Furthermore, even if the regions with different properties and optimal hyperparameters are mixed, the performance does not degrade much for any of the regions.
 
 Because of these advantages, an SDNN is a very practical function approximator suitable for various (but not very high-dimensional) tasks, including pattern classification. However, the implementation of an SDNN is difficult for general users because the pattern coding and selective desensitization operations are unique and not included in any existing library or development environment for machine learning. In addition, general users may face difficulty in writing an efficient program code because there exist many unpublished tricks to reduce the execution time greatly.
 
