### 2019-10-22 10c941b
Update paper.md
