### 2019-10-22 f23e52c
Update paper.md
