### 2019-10-24 6d99174
Update paper.md
