+++ doc/paper.md
@@ -61,14 +61,14 @@ In the following sections, we demonstrate the performance of the library by comp
 
 To demonstrate the performance of LIBSDNN and the SDNN itself, we perform a task of approximating a two-variable discontinuous function with various spatial frequencies depending on the area (Formula (1); Figure 1(a)); this was also shown in the study by Nonaka et al. [@Nonaka:2011]. The source code and the parameter files of this demonstration are provided in the LIBSDNN library package. Please check the libsdnn\\doc\\example folder.
 
-```math
+$$
 f(x) = \left\{
 \begin{array}{ll}
 1 & ((x-0.5)^2+(y-0.5)^2 <= 0.04) \\
 \frac{1+x}{2}sin^2(6\pixy^2) & (otherwise)
 \end{array}
 \right.
-```
+$$
 
 where $x,y \in [0,1]$.
 
