+++ doc/paper.md
@@ -65,7 +65,7 @@ $$
 f(x) = \left\{
 \begin{array}{ll}
 1 & ((x-0.5)^2+(y-0.5)^2 <= 0.04) \\
-\frac{1+x}{2}sin^{2}(6\pi xy^{2}) & (otherwise)
+\frac{1+x}{2}\sin^{2}(6\pi xy^{2}) & (otherwise)
 \end{array}
 \right.
 $$
