### 2019-11-17 dfe488c
Update paper.md
