### 2019-11-17 4627fcc
Update paper.md
