### 2019-11-17 a2305b3
Update paper.md
