### 2019-11-26 #1785 @arokem
@dirmeier, @HaoZeke : have you had a chance to take a look at this paper/software?
