### 2019-12-07 BIPL-HORIE/LIBSDNN#2 @HaoZeke
[JOSS] Fix missing DOIs 
Part of the [JOSS review here](https://github.com/openjournals/joss-reviews/issues/1785).

The following DOIs are missing as per @whedon:

```
- https://doi.org/10.1007/978-3-642-17534-3_6 may be missing for title: Practical Surface EMG Pattern Classification by Using a Selective Desensitization Neural Network
- https://doi.org/10.1109/scis-isis.2014.7044714 may be missing for title: Q-learning in continuous state-action space with redundant dimensions by using a selective desensitization neural network
```
