+++ doc/paper.bib
@@ -52,7 +52,7 @@ year={2010},
 publisher={Springer Berlin Heidelberg},
 address={Berlin, Heidelberg},
 pages={42-49},
-isbn={978-3-642-17534-3}
+isbn={978-3-642-17534-3},
 doi={10.1007/978-3-642-17534-3_6}
 }
 
@@ -72,7 +72,7 @@ doi = {10.2299/jsp.18.225}
   author={{Kobayashi}, T. and {Shibuya}, T. and {Morita}, M.},
   journal={2014 Joint 7th International Conference on Soft Computing and Intelligent Systems (SCIS) and 15th International Symposium on Advanced Intelligent Systems (ISIS)},
   year={2014},
-  pages={801-806}
+  pages={801-806},
   doi={10.1109/SCIS-ISIS.2014.7044714}
 }
 
