+++ doc/paper.bib
@@ -73,7 +73,7 @@ doi = {10.2299/jsp.18.225}
   journal={2014 Joint 7th International Conference on Soft Computing and Intelligent Systems (SCIS) and 15th International Symposium on Advanced Intelligent Systems (ISIS)},
   year={2014},
   pages={801-806},
-  doi={10.1109/SCIS-ISIS.2014.7044714}
+  doi = {10.1109/scis-isis.2014.7044714}
 }
 
 @article{Kobayashi:2015,
