### 2020-01-03 28a9a89
Delete parameter_file.txt
