### 2020-01-03 bdb7d8e
Update paper.md
