### 2020-01-20 #1785 @arokem
@BIPL-HORIE : without your response, we cannot move ahead with this paper. Please get back to me ASAP confirming that you have received these comments.
