### 2020-02-10 #1785 @HaoZeke
Do we have guidelines on rejection for delayed responses? To be totally fair I am well aware the reviews were not completed in the normal JOSS timeframe either..
