### 2020-02-10 #1785 @danielskatz
No, we generally try to get all submissions to the acceptable point
