### 2020-02-13 #1785 @BIPL-HORIE
Dear editor and reviewers,

I'm really sorry for my late response.
I regret submitting my paper even though I don't have enough time to revise.

Thank to the review, I found that my library does not have maintainability required as open source software.
And, now I'm willing to revise my library according to the review.
However, It will be a hard job for me and it cannot be fixed right away, considering my skills.

I'm really sorry about that. Now, I think that my submission should be rejected in this time.
I don't want to bother the editors and reviewers any more.

I would appreciate it if I could submit again when I finished revising the library.
